# bee_agents/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scholarship Analysis API",
    "description": "REST API for accessing scholarship application analysis data with multi-tenancy support",
    "version": "2.0.0"
  },
  "paths": {
...
    "/criteria": {
      "get": {
        "tags": [
          "Criteria"
        ],
        "summary": "List Criteria",
        "description": "List all available evaluation criteria files for a scholarship.\n\nReturns a list of all criteria files available for the specified scholarship,\nincluding their names, filenames, and fully qualified download URLs. This endpoint\nis useful for discovering what criteria types are available before requesting\nspecific criteria content.\n\n**Example Request:**\n```\nGET /criteria?scholarship=Delaney_Wings\n```\n\n**Example Response:**\n```json\n{\n  \"scholarship\": \"Delaney_Wings\",\n  \"criteria_count\": 5,\n  \"criteria\": [\n    {\n      \"name\": \"academic_criteria\",\n      \"filename\": \"academic_criteria.txt\",\n      \"url\": \"http://localhost:8200/criteria/Delaney_Wings/academic_criteria.txt\"\n    },\n    {\n      \"name\": \"application_criteria\",\n      \"filename\": \"application_criteria.txt\",\n      \"url\": \"http://localhost:8200/criteria/Delaney_Wings/application_criteria.txt\"\n    }\n  ]\n}\n```\n\nArgs:\n    scholarship: Name of the scholarship (e.g., 'Delaney_Wings' or 'Evans_Wings')\n    \nReturns:\n    Dictionary containing:\n    - scholarship: Name of the scholarship\n    - criteria_count: Number of criteria files found\n    - criteria: List of criteria objects with name, filename, and download URL\n    \nRaises:\n    HTTPException 404: If scholarship not found or no criteria directory exists",
        "operationId": "list_criteria",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
            "example": "Delaney_Wings"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/weights": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Scholarship Weights",
        "description": "Get current scoring weights for a scholarship from canonical config.",
        "operationId": "get_weights",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
            "example": "Delaney_Wings"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update Scholarship Weights",
        "description": "Update scoring weights in canonical config and regenerate artifacts.",
        "operationId": "update_weights",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
            "example": "Delaney_Wings"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/criteria": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Agent Criteria",
        "description": "Get current criteria text for an agent.",
        "operationId": "get_criteria",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
            "example": "Delaney_Wings"
          },
          {
            "name": "agent_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the agent (e.g., 'academic', 'essay', 'recommendation')",
              "title": "Agent Name"
            },
            "description": "Name of the agent (e.g., 'academic', 'essay', 'recommendation')",
            "example": "academic"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update Agent Criteria",
        "description": "Update criteria text for an agent and regenerate artifacts.",
        "operationId": "update_criteria",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
            "example": "Delaney_Wings"
          },
          {
            "name": "agent_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the agent (e.g., 'academic', 'essay', 'recommendation')",
              "title": "Agent Name"
            },
            "description": "Name of the agent (e.g., 'academic', 'essay', 'recommendation')",
            "example": "academic"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/criteria/regenerate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Regenerate Agent Criteria With Llm",
        "description": "Generate a new criteria draft for an agent using an LLM.\n\nThis endpoint does NOT persist changes automatically. It returns a\nproposed `criteria_text` that an admin can review and then apply via\nthe standard PUT /admin/{scholarship}/criteria/{agent_name} endpoint.",
        "operationId": "regen_criteria",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
            "example": "Delaney_Wings"
          },
          {
            "name": "agent_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name of the agent (e.g., 'academic', 'essay', 'recommendation')",
              "title": "Agent Name"
            },
            "description": "Name of the agent (e.g., 'academic', 'essay', 'recommendation')",
            "example": "academic"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/agents": {
      "get": {
        "tags": [
          "Configuration"
        ],
        "summary": "Get Agents Config",
        "description": "Get agent configuration and scoring weights for the scholarship.\n\nArgs:\n    scholarship: Name of the scholarship\n\nReturns:\n    Agent configuration including scoring weights and evaluation details",
        "operationId": "get_agents_config_agents_get",
        "parameters": [
          {
            "name": "scholarship",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')",
              "title": "Scholarship"
            },
            "description": "Scholarship name (e.g., 'Delaney_Wings' or 'Evans_Wings')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use 'Bearer &lt;token&gt;' from the chat /login endpoint."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8200",
      "description": "Development server (default port)"
    }
  ]
}